Move browse.js, remove sidebar, and normalize widths on browse index search vs results
